clean up app and nav buttons
